+++ backend/GarageControl.Core/ViewModels/Jobs/JobPartDetailsVM.cs
@@ -4,10 +4,10 @@ namespace GarageControl.Core.ViewModels.Jobs
     {
         public string PartId { get; set; } = null!;
         public string PartName { get; set; } = null!;
-        public double PlannedQuantity { get; set; }
-        public double SentQuantity { get; set; }
-        public double UsedQuantity { get; set; }
-        public double RequestedQuantity { get; set; }
+        public int PlannedQuantity { get; set; }
+        public int SentQuantity { get; set; }
+        public int UsedQuantity { get; set; }
+        public int RequestedQuantity { get; set; }
         public decimal Price { get; set; }
     }
 }
