+++ backend/GarageControl.Core/ViewModels/Parts/UpdatePartVM.cs
@@ -8,21 +8,17 @@ namespace GarageControl.Core.ViewModels.Parts
         [Required]
         [StringLength(PartConstants.NameMaxLength)]
         public string Name { get; set; } = null!;
-
         [Required]
         [StringLength(PartConstants.PartNumberMaxLength)]
         public string PartNumber { get; set; } = null!;
-
         [Required]
         [Range(0, double.MaxValue)]
         public decimal Price { get; set; }
-
         [Required]
-        [Range(0, double.MaxValue)]
-        public double Quantity { get; set; }
-
+        [Range(0, int.MaxValue)]
+        public int Quantity { get; set; }
         [Required]
-        [Range(0, double.MaxValue)]
-        public double MinimumQuantity { get; set; }
+        [Range(0, int.MaxValue)]
+        public int MinimumQuantity { get; set; }
     }
 }
